refactor(direct-control): extract notification dismissal source
Move the embedded App UI notification dismissal source into packages/civ7-direct-control/src/play/notifications/dismissal.ts and import it from the package facade. Keep wrapper-level polling and identity verification helpers in index.ts for the later composition pass.

Record the source-owner status in the support modularization OpenSpec workstream without marking the broader dismissal/verification atom complete. This is local package source relocation proof only, not runtime or live-game proof.

Validation: bun run --cwd packages/civ7-direct-control test -- notification-dismissal.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
