+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -114,6 +114,8 @@ runtime/direct-control claims.
     readback edge coverage to owning package suites before source extraction.
 - [ ] 4.3 Extract notification view/materialization atom.
 - [ ] 4.4 Extract notification dismissal/verification atom.
+  - [x] 4.4.1 Extract notification dismissal embedded source owner while
+        keeping wrapper-level polling and verification helpers in the facade.
 - [x] 4.5 Extract ready unit/city view atoms.
   - [x] 4.5.1 Extract unit move preview embedded source owner while keeping
         the public wrapper in the facade.
